split be/src/util/tdigest.h into headers and cpp files
