+++ be/src/util/CMakeLists.txt
@@ -109,6 +109,7 @@ set(UTIL_FILES
   gc_helper_smoothstep.cpp
   sha.cpp
   lru_cache.cpp
+  tdigest.cpp
 )
 
 # simdjson Runtime Implement Dispatch: https://github.com/simdjson/simdjson/blob/master/doc/implementation-selection.md#runtime-cpu-detection
